requirements.txt
http1==1.0.2
python-telegram-bot==13.9
discord-webhook